情景模式完成

+++ app/src/main/java/com/dsz/bean/AddNameBean.java
@@ -20,7 +20,7 @@ public class AddNameBean implements Serializable {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
